+++ packages/core/src/__tests__/qwik/convert-method-to-function.test.ts
@@ -1,7 +1,8 @@
+import { MethodMap } from 'src/generators/qwik/component-generator';
 import { convertMethodToFunction } from '../../generators/qwik/helpers/convert-method-to-function';
 
 describe('convertMethodToFunction', () => {
-  const methodMap: Record<string, 'method' | 'getter'> = {
+  const methodMap: MethodMap = {
     methodA: 'method',
     getterB: 'getter',
     getCssFromFont: 'method',
