+++ packages/core/src/generators/qwik/component-generator.ts
@@ -1,5 +1,4 @@
 import { format } from 'prettier/standalone';
-import { babelTransformExpression } from '../../helpers/babel-transform';
 import { fastClone } from '../../helpers/fast-clone';
 import { collectCss } from '../../helpers/styles/collect-css';
 import { MitosisComponent } from '../../types/mitosis-component';
@@ -16,34 +15,22 @@ import {
   runPreCodePlugins,
   runPreJsonPlugins,
 } from '../../modules/plugins';
-import traverse from 'traverse';
 import { stableInject } from './helpers/stable-inject';
 import { mergeOptions } from '../../helpers/merge-options';
+import {
+  emitStateMethodsAndRewriteBindings,
+  emitUseStore,
+  getLexicalScopeVars,
+  getStateMethodsAndGetters,
+  StateInit,
+} from './helpers/state';
+import { convertTypeScriptToJS } from './helpers/transform-code';
 
 Error.stackTraceLimit = 9999;
 
 const DEBUG = false;
 
 export interface ToQwikOptions extends BaseTranspilerOptions {}
-
-/**
- * Stores getters and initialization map.
- */
-type StateInit = [
-  StateValues,
-  /**
-   * Set of state initializers.
-   */
-  ...string[],
-];
-
-type PropertyName = string;
-type StateValue = string;
-
-/**
- * Map of getters that need to be rewritten to function invocations.
- */
-type StateValues = Record<PropertyName, StateValue>;
 
 const PLUGINS: Plugin[] = [
   () => ({
@@ -134,10 +121,8 @@ export const componentToQwik: TranspilerGenerator<ToQwikOptions> =
       emitStyles(file, css);
       DEBUG && file.exportConst('COMPONENT', JSON.stringify(component));
       let sourceFile = file.toString();
-      if (options.plugins) {
-        sourceFile = runPreCodePlugins(sourceFile, options.plugins);
-        sourceFile = runPostCodePlugins(sourceFile, options.plugins);
-      }
+      sourceFile = runPreCodePlugins(sourceFile, options.plugins);
+      sourceFile = runPostCodePlugins(sourceFile, options.plugins);
       return sourceFile;
     } catch (e) {
       console.error(e);
@@ -322,72 +307,10 @@ function emitStyles(file: File, css: string | null) {
   file.exportConst('STYLES', '`\n' + css.replace(/`/g, '\\`') + '`\n');
 }
 
-/**
- * @param file
- * @param stateInit
- */
-function emitUseStore(file: File, stateInit: StateInit) {
-  const state = stateInit[0];
-  const hasState = state && Object.keys(state).length > 0;
-  if (hasState) {
-    file.src.emit('const state=', file.import(file.qwikModule, 'useStore').localName);
-    if (file.options.isTypeScript) {
-      file.src.emit('<any>');
-    }
-    file.src.emit('(');
-    file.src.emit(stableInject(state));
-    file.src.emit(');');
-  } else {
-    // TODO hack for now so that `state` variable is defined, even though it is never read.
-    file.src.emit('const state' + (file.options.isTypeScript ? ': any' : '') + ' = {};');
-  }
-}
-
 function emitTypes(file: File, component: MitosisComponent) {
   if (file.options.isTypeScript) {
     component.types?.forEach((t) => file.src.emit(t, '\n'));
   }
-}
-
-function emitStateMethodsAndRewriteBindings(
-  file: File,
-  component: MitosisComponent,
-  metadata: Record<string, any>,
-): StateInit {
-  const lexicalArgs = getLexicalScopeVars(component);
-  const state: StateInit = emitStateMethods(file, component.state, lexicalArgs);
-  const methodMap = getStateMethodsAndGetters(component.state);
-  rewriteCodeExpr(component, methodMap, lexicalArgs, metadata.qwik?.replace);
-  return state;
-}
-
-const checkIsObjectWithCodeBlock = (obj: any): obj is { code: string } => {
-  return typeof obj == 'object' && obj?.code && typeof obj.code === 'string';
-};
-
-function rewriteCodeExpr(
-  component: MitosisComponent,
-  methodMap: Record<string, 'method' | 'getter'>,
-  lexicalArgs: string[],
-  replace: Record<string, string> | undefined = {},
-) {
-  traverse(component).forEach(function (item) {
-    if (!checkIsObjectWithCodeBlock(item)) {
-      return;
-    }
-
-    let code = convertMethodToFunction(item.code, methodMap, lexicalArgs);
-
-    Object.keys(replace).forEach((key) => {
-      code = code.replace(key, replace[key]);
-    });
-
-    item.code = code;
-  });
-}
-
-function getLexicalScopeVars(component: MitosisComponent) {
-  return ['props', 'state', ...Object.keys(component.refs), ...Object.keys(component.context.get)];
 }
 
 function emitImports(file: File, component: MitosisComponent) {
@@ -401,73 +324,8 @@ function emitImports(file: File, component: MitosisComponent) {
   });
 }
 
-function emitStateMethods(
-  file: File,
-  componentState: MitosisComponent['state'],
-  lexicalArgs: string[],
-): StateInit {
-  const stateValues: StateValues = {};
-  const stateInit: StateInit = [stateValues];
-  const methodMap = getStateMethodsAndGetters(componentState);
-  for (const key in componentState) {
-    const stateValue = componentState[key];
-
-    switch (stateValue?.type) {
-      case 'method':
-      case 'getter':
-      case 'function':
-        let code = stateValue.code;
-        let prefixIdx = 0;
-        if (stateValue.type === 'getter') {
-          prefixIdx += 'get '.length;
-        } else if (stateValue.type === 'function') {
-          prefixIdx += 'function '.length;
-        }
-        code = code.substring(prefixIdx);
-        code = convertMethodToFunction(code, methodMap, lexicalArgs).replace(
-          '(',
-          `(${lexicalArgs.join(',')},`,
-        );
-        const functionName = code.split(/\(/)[0];
-        if (stateValue.type === 'getter') {
-          stateInit.push(`state.${key}=${functionName}(${lexicalArgs.join(',')})`);
-        }
-        if (!file.options.isTypeScript) {
-          // Erase type information
-          code = convertTypeScriptToJS(code);
-        }
-        file.exportConst(functionName, 'function ' + code, true);
-
-        continue;
-
-      case 'property':
-        stateValues[key] = stateValue.code;
-        continue;
-    }
-  }
-
-  return stateInit;
-}
-
-function convertTypeScriptToJS(code: string): string {
-  return babelTransformExpression(code, {});
-}
-
 function extractGetterBody(code: string): string {
   const start = code.indexOf('{');
   const end = code.lastIndexOf('}');
   return code.substring(start + 1, end).trim();
-}
-
-function getStateMethodsAndGetters(
-  state: MitosisComponent['state'],
-): Record<string, 'method' | 'getter'> {
-  const methodMap: Record<string, 'method' | 'getter'> = {};
-  Object.keys(state).forEach((key) => {
-    const stateVal = state[key];
-    if (stateVal?.type === 'getter' || stateVal?.type === 'method') {
-      methodMap[key] = stateVal.type;
-    }
-  });
-  return methodMap;
 }
