+++ packages/core/src/generators/qwik/helpers/transform-code.ts
@@ -1,0 +1,5 @@
+import { babelTransformExpression } from 'src/helpers/babel-transform';
+
+export function convertTypeScriptToJS(code: string): string {
+  return babelTransformExpression(code, {});
+}
